# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Romanian Festival Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors (optionally filtered by year)
      security:
        - clerkAuth: []
      parameters:
        - name: yearId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vendors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a vendor by ID
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
  /vendors/{id}/review:
    patch:
      operationId: reviewVendor
      tags: [vendors]
      summary: Approve or reject a vendor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecision"
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}/final-approve:
    patch:
      operationId: finalApproveVendor
      tags: [vendors]
      summary: Final approve a vendor after payment
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Final approved vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}/assign:
    patch:
      operationId: assignVendorSpot
      tags: [vendors]
      summary: Assign spot number and location to a vendor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpotAssignment"
      responses:
        "200":
          description: Vendor with assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"

  # ─── SPONSORS ────────────────────────────────────────────
  /sponsors:
    get:
      operationId: listSponsors
      tags: [sponsors]
      summary: List sponsors (optionally filtered by year)
      security:
        - clerkAuth: []
      parameters:
        - name: yearId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sponsors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sponsor"
  /sponsors/{id}:
    get:
      operationId: getSponsor
      tags: [sponsors]
      summary: Get a sponsor by ID
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sponsor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
        "404":
          description: Not found
  /sponsors/{id}/review:
    patch:
      operationId: reviewSponsor
      tags: [sponsors]
      summary: Approve or reject a sponsor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecision"
      responses:
        "200":
          description: Updated sponsor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
  /sponsors/{id}/final-approve:
    patch:
      operationId: finalApproveSponsor
      tags: [sponsors]
      summary: Final approve a sponsor after payment
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Final approved sponsor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
  /sponsors/{id}/assign:
    patch:
      operationId: assignSponsorSpot
      tags: [sponsors]
      summary: Assign spot to a sponsor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpotAssignment"
      responses:
        "200":
          description: Sponsor with assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"

  # ─── VOLUNTEERS ──────────────────────────────────────────
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"

    post:
      operationId: inviteStaff
      tags: [staff]
      summary: Invite a new staff member
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInvite"
      responses:
        "201":
          description: Staff member invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    delete:
      operationId: removeStaff
      tags: [staff]
      summary: Remove a staff member
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ─── EXPORT ──────────────────────────────────────────────
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Vendor:
      type: object
      required: [id, yearId, name, businessName, email, phone, status, applicationData, createdAt]
      properties:
        id:
          type: integer
        yearId:
          type: integer
        name:
          type: string
        businessName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, payment_pending, paid, final_approved]
        applicationData:
          type: object
          additionalProperties: true
        agreementSigned:
          type: boolean
        spotNumber:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        reviewNote:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        finalApprovedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    Sponsor:
      type: object
      required: [id, yearId, name, orgName, email, phone, tier, status, applicationData, createdAt]
      properties:
        id:
          type: integer
        yearId:
          type: integer
        name:
          type: string
        orgName:
          type: string
        email:
          type: string
        phone:
          type: string
        tier:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, payment_pending, paid, final_approved]
        applicationData:
          type: object
          additionalProperties: true
        agreementSigned:
          type: boolean
        spotNumber:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        reviewNote:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        finalApprovedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    StaffMember:
      type: object
      required: [id, email, role, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: ["string", "null"]
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, staff]
        createdAt:
          type: string
...
    ReviewDecision:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        note:
          type: ["string", "null"]
...
    SpotAssignment:
      type: object
      required: [spotNumber, location]
      properties:
        spotNumber:
          type: string
        location:
          type: string
    StaffInvite:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, staff]