# lib/api-zod/src/generated/types/sponsor.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Romanian Festival Management API
 * OpenAPI spec version: 0.1.0
 */
import type { SponsorApplicationData } from './sponsorApplicationData';
import type { SponsorStatus } from './sponsorStatus';

export interface Sponsor {
  id: number;
  yearId: number;
  name: string;
  orgName: string;
  email: string;
  phone: string;
  tier: string;
  status: SponsorStatus;
  applicationData: SponsorApplicationData;
  agreementSigned?: boolean;
  /** @nullable */
  spotNumber?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  reviewNote?: string | null;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  finalApprovedAt?: string | null;
  createdAt: string;
}